wip-79095 BusinessProject: move generate invoice methods from InvoicingProjectService to ProjectGenerateInvoiceService
